Project Management Tool #HTML #CSS #JS
I'm thrilled to share that I've successfully completed my second task at Bharat Intern and what a fantastic journey.
creative this management tool using HTML ,CSS ,JS this project was a significant learning experience it's been a fantastic start to my internship journey. and I can't wait to tackle more exciting challenges in the days to come I want to express my gratitude to Bharat Intern for providing me with this incredible opportunity to learn and grow in a professional environment. I'm eager to take on new challenges and contribute my skills to the team stay tuned for more updates my projects as continue to explore the world of web development projects #Bharatintern #education #Bharatinternship